Separate form controls into different components
Created separate components for email, zipcode, and password
Used simple regex to display error messages to user if form control value
invalid
