+++ src/scenes/home/signup/signup.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import Form from 'shared/components/form/form';
-import DataFormItem from 'shared/components/form/dataFormItem/dataFormItem';
+import FormEmail from 'shared/components/form/formEmail/formEmail';
+import FormZipCode from 'shared/components/form/formZipCode/formZipCode';
 import FormPassword from 'shared/components/form/formPassword/formPassword';
 import styles from './signup.css';
 
@@ -9,8 +10,8 @@ class SignUp extends Component {
     return (
       <div className={styles.signup}>
         <Form>
-          <DataFormItem id="email" type="email" placeholder="Email" />
-          <DataFormItem id="zip" placeholder="Zip Code" />
+          <FormEmail id="email" placeholder="Email" />
+          <FormZipCode id="zip" placeholder="Zip Code" />
           <FormPassword id="password" placeholder="Password" />
         </Form>
       </div>
