--- src/shared/components/form/dataFormItem/dataFormItem.css
@@ -1,13 +1,0 @@
-.dataFormItem input {
-  background-color: #E6E6E6;
-  color: #000;
-  border: none;
-  font-size: 1.3em;
-  padding: 0 16px;
-  height: 50px;
-  width: 100%;
-}
-
-.dataFormItem input::placeholder  {
-    color: #000;
-}
