+++ src/shared/components/form/formEmail/formEmail.css
@@ -1,0 +1,3 @@
+.formEmail {
+  composes: formControl from '../form.css';
+}
