+++ src/shared/components/form/formPassword/formPassword.css
@@ -1,0 +1,3 @@
+.formPassword {
+  composes: formControl from '../form.css';
+}
