+++ src/shared/components/form/formZipCode/formZipCode.css
@@ -1,0 +1,3 @@
+.formZipCode {
+  composes: formControl from '../form.css';
+}
