+++ src/shared/components/form/formZipCode/formZipCode.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Label from 'shared/components/label/label';
-import styles from './formPassword.css';
+import styles from './formZipCode.css';
 
-class FormPassword extends Component {
+class FormZipCode extends Component {
   constructor() {
     super();
     this.state = {
@@ -19,7 +19,7 @@ class FormPassword extends Component {
 
   validate = (text) => {
     // todo: update regex if/when we decide to update password validation rules
-    const validationRegex = /^(?=.*[A-Z]).{6,}$/;
+    const validationRegex = /(^\d{5}$)|(^\d{5}-\d{4}$)/;  // US zip code regex from: http://stackoverflow.com/a/160583/4585393
 
     if (text.length > 0) {
       this.setState({ isValid: text.match(validationRegex) });
@@ -30,32 +30,33 @@ class FormPassword extends Component {
 
   render() {
     return (
-      <div className={styles.formPassword}>
+      <div className={styles.formZipCode}>
         {this.props.label && <Label htmlFor={this.props.id}>{this.props.displayName}</Label>}
         <input
+          autoComplete="off"
           id={this.props.id}
-          type="password"
+          type="text"
           value={this.state.text}
           placeholder={this.props.placeholder}
           onChange={this.handleChange}
         />
-        { !this.state.isValid && <span>Must be 6 characters long and incude a capitalized letter</span>}
+        { !this.state.isValid && <span>Must be a valid US zip code</span>}
       </div>
     );
   }
 }
 
-FormPassword.propTypes = {
+FormZipCode.propTypes = {
   label: PropTypes.bool,
   displayName: PropTypes.string,
   placeholder: PropTypes.string,
   id: PropTypes.string.isRequired
 };
 
-FormPassword.defaultProps = {
+FormZipCode.defaultProps = {
   displayName: null,
   label: true,
   placeholder: null
 };
 
-export default FormPassword;
+export default FormZipCode;
